fold performance recommendations into the overview tab
