+++ packages/react-devtools/src/components/RecommendationsList.module.scss
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2025 Palantir Technologies, Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+.list {
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+
+  // Blueprint minimal buttons default to dark text; keep the copy-prompt
+  // buttons readable on the panel and tinted recommendation cards.
+  :global(.bp6-button.bp6-minimal) {
+    color: var(--dt-text-secondary);
+  }
+}
+
+.head {
+  display: flex;
+  justify-content: flex-end;
+}
+
+.empty {
+  font-size: 12px;
+  opacity: 0.6;
+  padding: 8px 4px;
+}
+
+.recItem {
+  margin-bottom: 6px;
+}
+
+.recBody {
+  font-size: 12px;
+  margin-bottom: 6px;
+}
+
+.recActions {
+  display: flex;
+  justify-content: flex-end;
+}
